config: eslint homeFolder resolver
