+++ .eslintrc.json
@@ -20,6 +20,10 @@
   },
   "settings": {
     "import/resolver": {
+      "alias": {
+        "map": [["~", "./app/src"]],
+        "extensions": [".js", ".jsx", ".ts", ".tsx", ".json"]
+      },
       "node": {
         "extensions": [".js", ".jsx", ".ts", ".tsx", ".json"],
         "moduleDirectory": ["app/src", "node_modules"]
