+++ app/src/assets/scss/init.scss
@@ -19,6 +19,10 @@ input {
     @include t-title;
 }
 
+.t-title_2 {
+    @include t-title_2;
+}
+
 .t-center {
     display: block;
     text-align: center;
