+++ app/src/components/Leaderboard/Leaderboard.scss
@@ -1,0 +1,14 @@
+.leaderboard {
+    min-height: 348px;
+    max-height: 100%;
+    &__table {
+        width: 100%;
+        > table {
+            width: inherit;
+            min-width: 640px;
+        }
+    }
+    &__empty {
+        color: var(--color-second);
+    }
+}
