+++ app/src/components/LeaderboardRecord/LeaderboardRecord.scss
