+++ app/src/components/Menu/Menu.tsx
@@ -25,11 +25,8 @@ export const Menu = connector(
             <div className={`menu ${className}`}>
                 {isAuthenticated ? (
                     <>
-                        <Link to="/game" className="t-center mt-5">
+                        <Link to="/start" className="t-center mt-5">
                             <Button>New game</Button>
-                        </Link>
-                        <Link to="/leaderboard" className="t-center mt-5">
-                            <Button>Leaderboard</Button>
                         </Link>
                         <Link to="/user" className="t-center mt-5">
                             <Button>User info</Button>
