+++ app/src/pages/StartPage/StartPage.scss
