+++ app/src/pages/index.ts
@@ -1,7 +1,7 @@
 export * from './WelcomePage/WelcomePage';
 export * from './ErrorPage/ErrorPage';
+export * from './StartPage/StartPage';
 export * from './GamePage/GamePage';
-export * from './LeaderboardPage/LeaderboardPage';
 export * from './AuthorizationPage/AuthorizationPage';
 export * from './RegistrationPage/RegistrationPage';
 export * from './EditPage/EditPage';
