+++ app/src/router/AppRouter/AppRouter.tsx
@@ -1,15 +1,7 @@
 import React from 'react';
 import {BrowserRouter as Router, Switch, Route} from 'react-router-dom';
 import {PrivateRoute} from '../PrivateRoute/PrivateRoute';
-import {
-    ErrorPage,
-    GamePage,
-    LeaderboardPage,
-    AuthorizationPage,
-    RegistrationPage,
-    EditPage,
-    WelcomePage
-} from '../../pages';
+import {ErrorPage, StartPage, GamePage, AuthorizationPage, RegistrationPage, EditPage, WelcomePage} from '../../pages';
 
 /**
  * Роутер
@@ -30,11 +22,11 @@ export function AppRouter(): JSX.Element {
                 <PrivateRoute path="/user">
                     <EditPage />
                 </PrivateRoute>
+                <PrivateRoute path="/start">
+                    <StartPage />
+                </PrivateRoute>
                 <PrivateRoute path="/game">
                     <GamePage />
-                </PrivateRoute>
-                <PrivateRoute path="/leaderboard">
-                    <LeaderboardPage />
                 </PrivateRoute>
                 <Route path="*">
                     <ErrorPage type="404" />
