+++ package.json
@@ -74,6 +74,7 @@
     "eslint": "7.12.1",
     "eslint-config-airbnb": "18.2.0",
     "eslint-config-prettier": "6.15.0",
+    "eslint-import-resolver-alias": "1.1.2",
     "eslint-plugin-import": "2.22.1",
     "eslint-plugin-jsx-a11y": "6.4.1",
     "eslint-plugin-prettier": "3.1.4",
